Reçete stok türü hatası giderildi.

+++ TsiErp.Business/Entities/ProductionManagement/BillsofMaterial/Services/BillsofMaterialsAppService.cs
@@ -1,6 +1,7 @@
 ﻿using Microsoft.Extensions.Localization;
 using Tsi.Core.Aspects.Autofac.Caching;
 using Tsi.Core.Aspects.Autofac.Validation;
+using Tsi.Core.Utilities.EnumUtilities;
 using Tsi.Core.Utilities.ExceptionHandling.Exceptions;
 using Tsi.Core.Utilities.Results;
 using Tsi.Core.Utilities.Services.Business.ServiceRegistrations;
@@ -16,6 +17,7 @@ using TsiErp.Entities.Entities.ProductionManagement.BillsofMaterialLine;
 using TsiErp.Entities.Entities.ProductionManagement.BillsofMaterialLine.Dtos;
 using TsiErp.Entities.Entities.StockManagement.Product;
 using TsiErp.Entities.Entities.StockManagement.UnitSet;
+using TsiErp.Entities.Enums;
 using TsiErp.Entities.TableConstant;
 using TsiErp.Localizations.Resources.BillsofMaterials.Page;
 
@@ -88,7 +90,7 @@ namespace TsiErp.Business.Entities.BillsofMaterial.Services
                         LastModificationTime = null,
                         LastModifierId = Guid.Empty,
                         LineNr = item.LineNr,
-                        MaterialType = item.MaterialType,
+                        MaterialType =(int)item.MaterialType,
                         ProductID = item.ProductID,
                         Quantity = item.Quantity,
                         Size = item.Size,
@@ -334,7 +336,7 @@ namespace TsiErp.Business.Entities.BillsofMaterial.Services
                             LastModificationTime = null,
                             LastModifierId = Guid.Empty,
                             LineNr = item.LineNr,
-                            MaterialType = item.MaterialType,
+                            MaterialType = (int)item.MaterialType,
                             ProductID = item.ProductID,
                             Quantity = item.Quantity,
                             Size = item.Size,
@@ -367,7 +369,7 @@ namespace TsiErp.Business.Entities.BillsofMaterial.Services
                                 LastModificationTime = DateTime.Now,
                                 LastModifierId = LoginedUserService.UserId,
                                 LineNr = item.LineNr,
-                                MaterialType = item.MaterialType,
+                                MaterialType = (int)item.MaterialType,
                                 ProductID = item.ProductID,
                                 Quantity = item.Quantity,
                                 Size = item.Size,
